New version of the data structure.
I use a new version of the data structure. First there is a mapping of the folder by ID. And a mapping of the children for every folder.
Thanks to the new data structure, the folder chain is generated thanks to both mapping, generating a map of parents dynamically. With that version any update of the model will update the folder chain accordingly. Also I find it now to be the best of the two worlds between the flat data structure of chonky (the component is only aware of the current Files and an external state has to be kept externally in order to perform the update) and the full tree structure which lacks flexibility. Here depending on the case, a user can update only the content of a folder.